app/requirements.txt
asgiref==3.2.10
Django==3.0
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.6.1
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
zipp==3.1.0
django-cors-headers==3.4.0
